Generate show.config.php file from show.config - this is then included in ShowDataManager.
The old way was reading images.config and building array. Now we generate a new show.config.php every time we add to or update the show.config list (which drives the ordering).
Old: images.config
New: show.config

A problem was that images.config could not be ftp'ed after Apache PHP wrote it.